+++ fileCarver/recover.c
@@ -15,13 +15,178 @@
 // create a type of unsigned integers of length of 8 bits (1 byte)
 typedef uint8_t BYTE;
 
+char* checkFileType(int argc, char** argv, BYTE* block);
+
 // define blocksize to avoid magic number
 #define BLOCKSIZE 512
 
-int main (void)
+typedef enum {
+	PDF,
+	BMP,
+	JPG,
+	JPG2,
+	JPG3,
+	PNG,
+	GIF,
+	GIF2
+} fileType;
+
+// if ((block[0] == 0x25) && (block[1] == 0x50) && (block[2] == 0x44) && (block[3] == 0x46))
+// if ((block[0] == 0x42) && (block[1] == 0x4d))
+// if ((block[0] == 0xff) && (block[1] == 0xd8) && (block[2] == 0xff) && (block[3] == 0xe0 || block[3] == 0xe1))
+// if ((block[0] == 0x89) && (block[1] == 0x50) && (block[2] == 0x4e) && (block[3] == 0x47) && (block[4] == 0x0d) && (block[5] == 0x0a) && (block[6] == 0x1a) && (block[7] == 0x0a))
+
+BYTE* supportedFT[] = {	"\x25\x50\x44\x46",
+			"\x42\x4d",
+			"\xff\xd8\xff\xe0",
+			"\xff\xd8\xff\xe1",
+			"\xff\xd8\xff\xe8",
+			"\x89\x50\x4e\x47\x0d\x0a\x1a\x0a",
+			"\x47\x49\x46\x38\x37\x61",
+			"\x47\x49\x46\x38\x39\x61"
+			};
+int FTLens[] = { 4, 2, 4, 4, 4, 8, 6, 6};
+
+fileType convertArg(char* arg)
 {
-    // open memory card
-	FILE* fp = fopen("card.raw", "r");
+	if(strcmp(arg, "-pdf") == 0)
+		return PDF;
+	else if(strcmp(arg, "-bmp") == 0)
+		return BMP;
+	else if(strcmp(arg, "-jpg") == 0)
+		return JPG;
+	else if(strcmp(arg, "-png") == 0)
+		return PNG;
+	else if(strcmp(arg, "-gif") == 0)
+		return GIF;
+	else if(strcmp(arg, "-all") == 0)
+		return PDF;
+	else
+	{
+		printf("USAGE: ./recover <imageFile> <-options>\n");
+		printf(" -pdf		- Recovers pdf files\n");
+		printf(" -bmp		- Recovers bmp files\n");
+		printf(" -jpg		- Recovers jpg files\n");
+		printf(" -png		- Recovers png files\n");
+		printf(" -gif		- Recovers gif files\n");
+		exit(-1);
+	}	
+}
+
+char* checkFileType(int argc, char** argv, BYTE* block){
+
+	char* ret = NULL;	
+	
+	int i;
+	for(i = 2; i < argc; i++)
+	{
+		if(ret != NULL)
+			break;
+		int j;
+		BYTE* sig;
+		BYTE* sig2;
+		BYTE* sig3;
+		switch(convertArg(argv[i]))
+		{
+			case PDF: ;
+				sig = supportedFT[PDF];
+				for(j = 0; j < FTLens[PDF]; j++)
+				{
+					if(block[j] != sig[j])
+					{
+						ret = NULL;
+						break;
+					}else{
+						ret = ".pdf";
+					}
+				}		
+				if(strcmp(argv[2], "-all") != 0 || ret != NULL)
+					break;
+
+			case BMP: ;
+				sig = supportedFT[BMP];
+				for(j = 0; j < FTLens[BMP]; j++)
+				{
+					if(block[j] != sig[j])
+					{
+						ret = NULL;
+						break;
+					}else{
+						ret = ".bmp";
+					}
+				}
+				if(strcmp(argv[2], "-all") != 0 || ret != NULL)
+					break;
+				
+			case JPG: ;
+				sig = supportedFT[JPG];
+				sig2 = supportedFT[JPG2];
+				sig3 = supportedFT[JPG3];
+				for(j = 0; j < FTLens[JPG]; j++)
+				{
+					if((block[j] != sig[j]) && (block[j] != sig2[j]) && (block[j] != sig3[j]))
+					{	
+						ret = NULL;
+						break;
+					}else{
+						ret = ".jpg";
+					}
+				}
+				if(strcmp(argv[2], "-all") != 0 || ret != NULL)
+					break;
+				
+			case PNG: ;
+				sig = supportedFT[PNG];
+				for(j = 0; j < FTLens[PNG]; j++)
+				{
+					if(block[j] != sig[j])
+					{
+						ret = NULL;
+						break;
+					}else{
+						ret = ".png";
+					}
+				}
+				if(strcmp(argv[2], "-all") != 0 || ret != NULL)
+					break;
+				
+			case GIF: ;
+				sig = supportedFT[GIF];
+				sig2 = supportedFT[GIF2];
+				for(j = 0; j < FTLens[GIF]; j++)
+				{
+					if( (block[j] != sig[j]) && (block[j] != sig2[j]))
+					{
+						ret = NULL;
+						break;
+					}else{
+						ret = ".gif";
+					}
+				}
+				break;
+
+			default:
+				ret = NULL;
+		}
+	}
+	return ret;
+}
+
+int main (int argc, char** argv)
+{
+	// Check arguments for usage
+	if(argc <= 2)
+	{
+		printf("USAGE: ./recover <imageFile> [-options]\n");
+		printf(" -pdf		- Recovers pdf files\n");
+		printf(" -bmp		- Recovers bmp files\n");
+		printf(" -jpg		- Recovers jpg files\n");
+		printf(" -png		- Recovers png files\n");
+		exit(-1);
+	} 
+
+    // open specified file 
+	FILE* fp = fopen(argv[1], "r");
 	
 	// if fp is NULL (non-existent file), return 1
     if (fp == NULL)
@@ -35,7 +200,7 @@ int main (void)
     outfile = NULL;
     
     // create a variable to name the new image files
-    int jpgnum = 0;
+    int fileNum = 0;
     
     // forever loop that will end when the file has been read
     while (1)
@@ -44,7 +209,8 @@ int main (void)
         BYTE block[BLOCKSIZE] = {};
         
         // iterate over each byte in each block until you reach the end of the file
-        for (int i = 0; i < BLOCKSIZE; i++)
+	int i;
+        for (i = 0; i < BLOCKSIZE; i++)
         {
             // if you reach the end of file, close the file
             if (feof(fp))
@@ -61,10 +227,13 @@ int main (void)
             // read one byte at a time
             fread(&block[i], sizeof (BYTE), 1, fp);
         }
+
+		char* ext = NULL;
         
+		ext = checkFileType(argc, argv, block);
         // if block's first 4 bytes match those of a jpg (start of new jpg)      
-        if ((block[0] == 0xff) && (block[1] == 0xd8) && (block[2] == 0xff) && (block[3] == 0xe0 || block[3] == 0xe1))
-        {
+		if (ext != NULL)
+		{
             // close previously opened output file pointer if necessary
             if (outfile != NULL)
             {
@@ -75,8 +244,9 @@ int main (void)
             char filename[4];
             
             // name file using ###.jpg format
-            sprintf(filename, "%03d.jpg", jpgnum);
-            jpgnum++; 
+            sprintf(filename, "%03d%s", fileNum, ext);
+			printf("file %d = %s\n", fileNum, filename);
+            fileNum++; 
             
             // open the output file pointer and make sure it's not NULL
             if ((outfile = fopen(filename, "w")) == NULL)
